+++ app/src/main/java/com/app/mathracer/data/repository/FriendRepository.kt
@@ -1,4 +1,4 @@
-package com.app.mathracer.data.repositories
+package com.app.mathracer.data.repository
 
 import com.app.mathracer.data.model.Friend
 import android.util.Log
