+++ app/src/main/java/com/app/mathracer/data/repository/SoloGameRepositoryImpl.kt
@@ -1,15 +1,19 @@
-package com.app.mathracer.data.repositories
+package com.app.mathracer.data.repository
 
+import android.util.Log
 import com.app.mathracer.data.model.SoloAnswerResponse
 import com.app.mathracer.data.model.SoloGameStartResponse
 import com.app.mathracer.data.model.SoloGameUpdateResponse
+import com.app.mathracer.data.model.WildCard
 import com.app.mathracer.data.network.RetrofitClient
 import com.app.mathracer.domain.repositories.SoloGameRepository
+import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.isActive
 import retrofit2.HttpException
 import java.io.IOException
+import kotlin.coroutines.coroutineContext
 
 class SoloGameRepositoryImpl : SoloGameRepository {
     
@@ -92,7 +96,7 @@ class SoloGameRepositoryImpl : SoloGameRepository {
                 once.fold(
                     onSuccess = { emit(it) },
                     onFailure = {
-                        android.util.Log.e("SoloGameRepository", "Error one-shot game update", it)
+                        Log.e("SoloGameRepository", "Error one-shot game update", it)
                         emit(null)
                     }
                 )
@@ -101,7 +105,7 @@ class SoloGameRepositoryImpl : SoloGameRepository {
 
             // polling continuo
             var backoffMs = intervalMs // podés ajustar backoff si falla
-            while (kotlin.coroutines.coroutineContext.isActive) {
+            while (coroutineContext.isActive) {
                 val result = getSoloGameUpdate(gameId)
                 result.fold(
                     onSuccess = { update ->
@@ -109,18 +113,18 @@ class SoloGameRepositoryImpl : SoloGameRepository {
                         backoffMs = intervalMs // reset backoff al éxito
                     },
                     onFailure = { exception ->
-                        android.util.Log.e("SoloGameRepository", "Error polling game update", exception)
+                        Log.e("SoloGameRepository", "Error polling game update", exception)
                         emit(null)
                         // backoff simple opcional: no lo hagas exponencial si no querés
                         backoffMs = (backoffMs.coerceAtMost(60_000L))
                     }
                 )
-                kotlinx.coroutines.delay(backoffMs)
+                delay(backoffMs)
             }
         }
     }
 
-    override suspend fun useWildcard(gameId: Int, wildcardId: Int): Result<com.app.mathracer.data.model.WildCard> {
+    override suspend fun useWildcard(gameId: Int, wildcardId: Int): Result<WildCard> {
         return try {
             val token = UserRemoteRepository.getIdToken()
             val header = token?.let { "Bearer $it" }
