+++ app/src/main/java/com/app/mathracer/ui/screens/home/viewmodel/HomeViewModel.kt
@@ -2,7 +2,7 @@ package com.app.mathracer.ui.screens.home.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.app.mathracer.data.repositories.UserRemoteRepository
+import com.app.mathracer.data.repository.UserRemoteRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.delay
