+++ app/src/main/java/com/app/mathracer/ui/screens/login/viewmodel/LoginViewModel.kt
@@ -11,7 +11,7 @@ import com.google.android.gms.auth.api.signin.GoogleSignIn
 import com.google.android.gms.common.api.ApiException
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.GoogleAuthProvider
-import com.app.mathracer.data.repositories.UserRemoteRepository
+import com.app.mathracer.data.repository.UserRemoteRepository
 import com.app.mathracer.data.CurrentUser
 import com.app.mathracer.data.model.User
 import androidx.lifecycle.viewModelScope
