+++ app/src/main/java/com/app/mathracer/ui/screens/worlds/viewmodel/WorldsViewModel.kt
@@ -4,9 +4,9 @@ import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.app.mathracer.data.model.WorldDto
-import com.app.mathracer.data.repositories.LevelsRemoteRepository
+import com.app.mathracer.data.repository.LevelsRemoteRepository
 import com.app.mathracer.data.model.Worlds
-import com.app.mathracer.data.repositories.WorldsRemoteRepository
+import com.app.mathracer.data.repository.WorldsRemoteRepository
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
